Add guard for controller click vs user click
The guard is predicated on the fact that Hulu specifically has a
different node that actually gets clicked vs the controller that I have
set.  I have no idea if this is true for the other services that I want
to support, but eventually I'll find out I guess.

This should work to avoid the whole back and forth of pause play.
Programmatic clicks can be separated from user generated clicks.  I need
to figure out a better way to handle that though.