Add specialized event dispatch
I finally ran across the docs I needed to put 2 and 2 together.

By creating my own even instance I can modify the properties, and add
one that will only exist on introspection for events propogated from
server actions.  This will make it much cleaner to gaurd against user vs
server events.

This is important because the other gaurd required that I set up the
controller querySelector and that it be different from the overlay that
the user clicked.  This may not always be the case, and is a lot of
overhead on my part.

By gaurding the way I do now, I can create a context option, similar to
uBlock, where the user sets the controller by right clicking it.  The
location of the page is introspected and the domain is used to store the
controller to localstorage.

The only thing left to do now is implement it.  Lol